Template: move document generation to a common place
The markdown documentation generation in the Template class is rather
common and makes more sense to put it in the Check class so it can be
used by other systems.
