+++ src/dct/templates/Template.lua
@@ -6,7 +6,6 @@ local class    = require("libs.namedclass")
 local check    = require("libs.check")
 local utils    = require("libs.utils")
 local dctenum  = require("dct.enum")
-local dctutils = require("dct.libs.utils")
 local STM      = require("dct.templates.STM")
 local Agent    = require("dct.assets.Agent")
 local Logger   = dct.Logger.getByName("Template")
@@ -189,132 +188,11 @@ function Template.fromFile(dctfile, stmfile)
 	return Template(template)
 end
 
-local Checker = require("dct.templates.checkers.Check")
-
-local value_header = {
-	[Checker.valuetype.VALUES]    = "specific values",
-	[Checker.valuetype.INT]       = "number",
-	[Checker.valuetype.RANGE]     = "range",
-	[Checker.valuetype.STRING]    = "string",
-	[Checker.valuetype.BOOL]      = "boolean (true/false)",
-	[Checker.valuetype.TABLEKEYS] = "specific values",
-	[Checker.valuetype.TABLE]     = "table",
-}
-
-local function is_required(option)
-	local s = " - _required:_ "
-
-	if option.default ~= nil then
-		s = s.."no"
-
-		if not (option.default == "") and
-		   type(option.default) ~= "table" and
-		   option.type ~= Checker.valuetype.VALUES then
-			s = s.."\n - _default:_ "..tostring(option.default)
-		elseif option.type == Checker.valuetype.VALUES then
-			local found = nil
-			for key, data in pairs(option.values) do
-				if data.value == option.default then
-					found = key
-					break
-				end
-			end
-
-			if found ~= nil then
-				s = s.."\n - _default:_ "..tostring(found)
-			end
-		end
-	else
-		s = s.."yes"
-	end
-	return s
-end
-
-local function option_summary(option)
-	local summary = is_required(option).."\n"
-
-	summary = summary.." - _value:_ "..value_header[option.type]
-	if option.type == Checker.valuetype.RANGE then
-		summary = summary..string.format(" [%d, %d]",
-			option.values[1], option.values[2])
-	end
-	summary = summary.."\n"
-	if option.agent then
-		summary = summary.." - _agent:_ true\n"
-	end
-	if option.deprecated then
-		summary = summary.."\n_NOTE: this option has been "..
-			  "deprecated._\n"
-	end
-	return summary
-end
-
-local function option_description(option)
-	local desc = option.description.."\n"
-
-	if option.type == Checker.valuetype.VALUES or
-	   option.type == Checker.valuetype.TABLEKEYS or
-	   (option.type == Checker.valuetype.TABLE and
-	    option.values ~= nil) then
-		local values = ""
-		for k, v in utils.sortedpairs(option.values) do
-			values = values.." - `"..k.."`"
-			if option.type == Checker.valuetype.VALUES then
-				values = values.." - "..v.description
-			end
-			values = values.."\n"
-		end
-		local len = string.len(values)
-		values = string.sub(values, 1, len - 1)
-
-		desc = dctutils.interp(desc, {
-			["VALUES"] = values,
-		})
-	end
-	return desc
-end
-
-local function write_section(level, name, data)
-	if next(data.options) == nil then
-		return
-	end
-
-	print(string.format("\n%s %s\n", string.rep("#", level), name))
-	if data.description then
-		print(data.description)
-	end
-
-	for optname, optdata in utils.sortedpairs(data.options) do
-		print(string.format("\n%s `%s`\n",
-				    string.rep("#", level+1), optname))
-		print(option_summary(optdata))
-		print(option_description(optdata))
-	end
-end
-
 --- class function to generate Template documentation.
 -- Generate markdown styled documentation for all options a campaign
 -- designer can use to specify a template.
 function Template.genDocs()
-	local sections = {}
-	for _, c in pairs(checkers) do
-		local doc = c:doc()
-		if sections[doc.section] == nil and next(doc.options) then
-			sections[doc.section] = {}
-			sections[doc.section]["options"] = {}
-		end
-
-		if doc.description then
-			sections[doc.section]["description"] = doc.description
-		end
-
-		for key, val in pairs(doc.options) do
-			if val.nodoc ~= true then
-				sections[doc.section]["options"][key] = val
-			end
-		end
-	end
-
+	local Checker = require("dct.templates.checkers.Check")
 	local header = [[
 # Template Attributes
 
@@ -335,10 +213,7 @@ reasonable defaults based on template type, composition, and other factors
 will be considered when selecting the default.
 ]]
 
-	print(header)
-	for name, data in utils.sortedpairs(sections) do
-		write_section(2, name, data)
-	end
+	Checker.genDocs(header, checkers)
 end
 
 --- Create a DCT game object from the template definition.
